requirements.txt
fastapi
uvicorn[standard]
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
sqlalchemy
pydantic
aiofiles
psycopg2-binary